Refactor POSIX semaphore header structure
Move from a single header with a massive ifdef section into two separate headers.
